requirements.txt
pandas
numpy
openpyxl
pyyaml
scikit-learn
xgboost
shap
matplotlib
scipy
pyarrow
ipywidgets
scikit-learn
catboost